Move `FnCtxt::get_fn_decl` to `TyCtxt`
