+++ compiler/rustc_hir_typeck/src/fn_ctxt/_impl.rs
@@ -9,7 +9,7 @@ use rustc_hir as hir;
 use rustc_hir::def::{CtorOf, DefKind, Res};
 use rustc_hir::def_id::DefId;
 use rustc_hir::lang_items::LangItem;
-use rustc_hir::{ExprKind, GenericArg, HirId, Node, QPath};
+use rustc_hir::{ExprKind, GenericArg, HirId, QPath};
 use rustc_hir_analysis::hir_ty_lowering::errors::GenericsArgsErrExtend;
 use rustc_hir_analysis::hir_ty_lowering::generics::{
     check_generic_arg_count_for_call, lower_generic_args,
@@ -34,7 +34,7 @@ use rustc_middle::{bug, span_bug};
 use rustc_session::lint;
 use rustc_span::def_id::LocalDefId;
 use rustc_span::hygiene::DesugaringKind;
-use rustc_span::symbol::{kw, sym};
+use rustc_span::symbol::kw;
 use rustc_span::Span;
 use rustc_target::abi::FieldIdx;
 use rustc_trait_selection::traits::error_reporting::TypeErrCtxtExt as _;
@@ -863,84 +863,6 @@ impl<'a, 'tcx> FnCtxt<'a, 'tcx> {
             Some(ty),
             slice::from_ref(&**item_segment),
         )
-    }
-
-    /// Given a `HirId`, return the `HirId` of the enclosing function, its `FnDecl`, and whether a
-    /// suggestion can be made, `None` otherwise.
-    pub fn get_fn_decl(
-        &self,
-        blk_id: HirId,
-    ) -> Option<(LocalDefId, &'tcx hir::FnDecl<'tcx>, bool)> {
-        // Get enclosing Fn, if it is a function or a trait method, unless there's a `loop` or
-        // `while` before reaching it, as block tail returns are not available in them.
-        self.tcx.hir().get_fn_id_for_return_block(blk_id).and_then(|item_id| {
-            match self.tcx.hir_node(item_id) {
-                Node::Item(&hir::Item {
-                    ident,
-                    kind: hir::ItemKind::Fn(ref sig, ..),
-                    owner_id,
-                    ..
-                }) => {
-                    // This is less than ideal, it will not suggest a return type span on any
-                    // method called `main`, regardless of whether it is actually the entry point,
-                    // but it will still present it as the reason for the expected type.
-                    Some((owner_id.def_id, sig.decl, ident.name != sym::main))
-                }
-                Node::TraitItem(&hir::TraitItem {
-                    kind: hir::TraitItemKind::Fn(ref sig, ..),
-                    owner_id,
-                    ..
-                }) => Some((owner_id.def_id, sig.decl, true)),
-                // FIXME: Suggestable if this is not a trait implementation
-                Node::ImplItem(&hir::ImplItem {
-                    kind: hir::ImplItemKind::Fn(ref sig, ..),
-                    owner_id,
-                    ..
-                }) => Some((owner_id.def_id, sig.decl, false)),
-                Node::Expr(&hir::Expr {
-                    hir_id,
-                    kind: hir::ExprKind::Closure(&hir::Closure { def_id, kind, fn_decl, .. }),
-                    ..
-                }) => {
-                    match kind {
-                        hir::ClosureKind::CoroutineClosure(_) => {
-                            // FIXME(async_closures): Implement this.
-                            return None;
-                        }
-                        hir::ClosureKind::Closure => Some((def_id, fn_decl, true)),
-                        hir::ClosureKind::Coroutine(hir::CoroutineKind::Desugared(
-                            _,
-                            hir::CoroutineSource::Fn,
-                        )) => {
-                            let (ident, sig, owner_id) = match self.tcx.parent_hir_node(hir_id) {
-                                Node::Item(&hir::Item {
-                                    ident,
-                                    kind: hir::ItemKind::Fn(ref sig, ..),
-                                    owner_id,
-                                    ..
-                                }) => (ident, sig, owner_id),
-                                Node::TraitItem(&hir::TraitItem {
-                                    ident,
-                                    kind: hir::TraitItemKind::Fn(ref sig, ..),
-                                    owner_id,
-                                    ..
-                                }) => (ident, sig, owner_id),
-                                Node::ImplItem(&hir::ImplItem {
-                                    ident,
-                                    kind: hir::ImplItemKind::Fn(ref sig, ..),
-                                    owner_id,
-                                    ..
-                                }) => (ident, sig, owner_id),
-                                _ => return None,
-                            };
-                            Some((owner_id.def_id, sig.decl, ident.name != sym::main))
-                        }
-                        _ => None,
-                    }
-                }
-                _ => None,
-            }
-        })
     }
 
     pub(crate) fn note_internal_mutation_in_method(
